Platfom verify

+++ application_service/app_service.cc
@@ -677,7 +677,7 @@ done:
     }
     string str_resp;
     if (resp.SerializeToString(&str_resp)) {
-      if (sized_socket_write(client, str_resp.size(), (byte*)str_resp.data()) < str_resp.size()) {
+      if (sized_socket_write(client, str_resp.size(), (byte*)str_resp.data()) < (int)str_resp.size()) {
         printf("Write failed\n");
       }
     }
